backend/requirements.txt
flask
flask-cors
PyMuPDF
python-docx
werkzeug
